Inline `.Protocol.State.*` into `.Protocol`
This will be required for the next commit.
