+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -68,7 +68,6 @@ library
                        Ouroboros.Consensus.HardFork.Combinator.Protocol
                        Ouroboros.Consensus.HardFork.Combinator.Protocol.ChainSel
                        Ouroboros.Consensus.HardFork.Combinator.Protocol.LedgerView
-                       Ouroboros.Consensus.HardFork.Combinator.Protocol.State
                        Ouroboros.Consensus.HardFork.Combinator.State
                        Ouroboros.Consensus.HardFork.Combinator.State.Infra
                        Ouroboros.Consensus.HardFork.Combinator.Translation
